Добавил кнопочку отбора по рейтингу
